+++ src/components/KeyShortCutsWrapper/KeyShortCutsWrapper.test.js
@@ -1,8 +1,8 @@
 import {
   render,
   cleanup,
-} from '@testing-library/react';
-import '@testing-library/jest-dom/extend-expect';
+} from '@folio/jest-config-stripes/testing-library/react';
+import '@folio/jest-config-stripes/testing-library/jest-dom';
 
 import {
   CommandList,