+++ src/views/AuthoritiesSearch/AuthoritiesSearch.test.js
@@ -2,7 +2,7 @@ import {
   waitFor,
   render,
   fireEvent,
-} from '@testing-library/react';
+} from '@folio/jest-config-stripes/testing-library/react';
 
 import { runAxeTest } from '@folio/stripes-testing';
 import { searchResultListColumns } from '@folio/stripes-authority-components';