+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [UIMARCAUTH-289](https://issues.folio.org/browse/UIMARCAUTH-289) New permission: MARC authority: Create MARC authority record.
 - [UIMARCAUTH-290](https://issues.folio.org/browse/UIMARCAUTH-290) Create Orig Authority Record: Add MARC authority action.
 - [UIMARCAUTH-291](https://issues.folio.org/browse/UIMARCAUTH-291) Update markHighlightedFields function with authority Mapping Rules. Added new permission to get mapping-rules.
+- [UIMARCAUTH-300](https://issues.folio.org/browse/UIMARCAUTH-300) *BREAKING* Bump `react` to `v18`.
 
 ## [3.0.2](https://github.com/folio-org/ui-marc-authorities/tree/v3.0.2) (2023-03-24)